Add `scroll-behavior: smooth;` to main.css (#461)
This commit is rather simple. It simply makes the scrolling when you scroll by id much smoother. For example, when you use the Back to top button in the footer, it won't jump up so quickly, it'll instead scroll. This is useful because it will surprise people less and it'll seem cleaner and less choppy too.