Clarify distinction between useSession & useCurrentUser
Rereading through the docs it's unclear what the use-case difference is between `useSession()` and `useCurrentUser()`.

This removes the mentioning of suppressing the first render since it's unrelated to the useSession vs useCurrentUser discussion. 

I still think we need a *reason* that you would use useCurrentUser. I'm not really sure what that would be though?