docs(multitenancy): fix example update mutation
In the current example, the code snippet is invalid. `update()` is limited to unique model fields, so the only way to execute this call is to add a unique constraint on `id` and `organizationId`. There's also `updateMany()`, but I wouldn't use that as an example as it will scan all table rows and can hurt performance. To my knowledge Prisma doesn't yet support limiting `updateMany()` to the first updated record.